small fix to DvQuantityIndex

+++ grails-app/domain/ehr/clinical_documents/data/DvQuantityIndex.groovy
@@ -3,6 +3,6 @@ package ehr.clinical_documents.data
 class DvQuantityIndex extends DataValueIndex {
    
    // data
-   float magnitude
+   double magnitude
    String units
 }
